Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 
 int main(){
-    float num;
-    scanf("%f",&num);
-    printf("Area: %.2f",(22/7*num*num));
+    int num;
+    scanf("%d",&num);
+    printf("Area: %.2f",(float)(22/7*num*num));
 }
